Moved population/ingest functionality to a seperate structure
