+++ nucleus/logger.py
@@ -1,0 +1,9 @@
+import logging
+
+import requests
+
+logger = logging.getLogger(__name__)
+logging.basicConfig()
+logging.getLogger(requests.packages.urllib3.__package__).setLevel(
+    logging.ERROR
+)
