+++ nucleus/retry_strategy.py
@@ -1,0 +1,3 @@
+class RetryStrategy:
+    statuses = {503, 504}
+    sleep_times = [1, 3, 9]
